requirements.txt
# --- Base ---
pandas~=2.3.3
numpy~=2.3.4
langchain
faiss-cpu~=1.12.0
networkx~=3.5
sentence-transformers~=5.1.2
openai
rouge-score
nltk~=3.9.2
pyarrow
fastparquet
rank_bm25~=0.2.2
pyyaml~=6.0.3
dotenv~=0.9.9
tqdm~=4.67.1
google-genai

# --- PDF and text processing ---
PyPDF2~=3.0.1
langdetect~=1.0.9
pytesseract
pdf2image
pillow

# --- Jupyter and ipynb ---
jupyter
notebook
ipykernel
nbformat
nbconvert

# --- Optional utilities ---
tika
beautifulsoup4
pdfkit
python-dotenv~=1.2.1